requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Data & Utilities
kaggle>=1.6.0
Pillow>=10.0.0
numpy>=1.24.0

# Visualization (optional but recommended)
matplotlib>=3.7.0

# Jupyter (optional — for notebook workflow)
notebook>=7.0.0
ipywidgets>=8.0.0
